chore: move UniqueStrList to constraints.py
While we're at it, make it generic :-)
